Add org-mode composing to mu4e emails
With this config, mu4e emails can be written in org-mode, which is later
converted to HTML  and sent like formatted text. It  works like a wonder
but we have to  set some shady keybindings to make  it work, since there
is  no   way  to  automatically   activate  the  mode   when  composing.
mu4e-compose-mode-hook results in  a nasty loop that  prevents any email
from being sent. The best option would  be to enable a function to set a
custom function that activates message composing.