Include initial code for not using emacs-doom
This commit performs fresh start for the configuration after roughly a
year of using emacs-doom. It is heavily inspired in the setup I use
with it, and it will evolve to shamelessly copy some of the features
and bindings. This, however, aims to be a fresh start to use a less
bloated Emacs, loading the minimum packages possible and allowing for
more stability (due to less moving parts). It is also a good learning
exercise, that forces me to learn how some things are implemented in
emacs-doom.